+++ src/model/mod.rs
@@ -8,6 +8,7 @@ mod compatibility;
 mod contract;
 mod control_flow;
 mod data_flow;
+mod endpoints;
 mod execution;
 mod extension;
 mod extension_value;
@@ -30,6 +31,9 @@ pub use compatibility::*;
 pub use contract::*;
 pub use control_flow::*;
 pub use data_flow::*;
+pub use endpoints::{
+    data_flow_endpoint_known, data_flow_step_dependency, known_data_flow_endpoints,
+};
 pub use execution::*;
 pub use extension::*;
 pub use extension_value::{ExtensionMap, ExtensionValue};